requirements.txt
torch>=1.9.0,<2.0.0
torchaudio>=0.9.0,<2.0.0
numpy>=1.19.2,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=0.24.2,<1.0.0
tqdm>=4.62.2,<5.0.0
PyYAML>=5.4.1,<6.0.0
tensorboard>=2.4.0,<3.0.0
click>=8.0.0,<9.0.0
soundfile>=0.10.3,<1.0.0
librosa>=0.8.1,<1.0.0
scipy>=1.7.0,<2.0.0
matplotlib>=3.4.0,<4.0.0
